+++ ReleaseNotes.md
@@ -54,3 +54,4 @@
 - Add MileGetFileSizeByHandle function.
 - Add MileGetFileAllocationSizeByHandle function.
 - Add MileGetCompressedFileSizeByHandle function.
+- Add Mile::CreateThread template function.